refactor(cli): split run output tests
